+++ packages/plugin-miniapp/src/miniapp/index.ts
@@ -34,7 +34,6 @@ const getMiniappTask = ({
   command,
   target,
   configAPI,
-  dataCache,
   runtimeDir,
   nativeConfig,
 }): Config => {
@@ -119,17 +118,6 @@ const getMiniappTask = ({
     enableCopyPlugin: isPublicDirExist, // Only when public dir exists should copy-webpack-plugin be enabled
     swcOptions: {
       removeExportExprs: ['getServerData', 'getStaticData'],
-      getRoutePaths: () => {
-        const routes = dataCache.get('routes');
-
-        const routeManifest = JSON.parse(routes)?.routeManifest || {};
-        const routeFiles = Object.keys(routeManifest).map((key) => {
-          const { file } = routeManifest[key];
-          return `src/pages/${file}`;
-        });
-
-        return routeFiles;
-      },
     },
     cssFilename: `[name]${fileType.style}`,
     cssChunkFilename: `[name]${fileType.style}`,