+++ packages/runtime/src/RouteWrapper.tsx
@@ -1,15 +1,16 @@
 import * as React from 'react';
-import type { RouteWrapperConfig } from './types.js';
+import type { RouteWrapperConfig, ComponentModule } from './types.js';
 
 interface Props {
   id: string;
   isLayout?: boolean;
   wrappers?: RouteWrapperConfig[];
   children?: React.ReactNode;
+  routeExports: ComponentModule;
 }
 
 export default function RouteWrapper(props: Props) {
-  const { wrappers = [], id, isLayout } = props;
+  const { wrappers = [], id, isLayout, routeExports } = props;
   // layout should only be wrapped by Wrapper with `layout: true`
   const filtered = isLayout ? wrappers.filter(wrapper => wrapper.layout === true) : wrappers;
   const RouteWrappers = filtered.map(item => item.Wrapper);
@@ -18,7 +19,7 @@ export default function RouteWrapper(props: Props) {
 
   if (RouteWrappers.length) {
     element = RouteWrappers.reduce((preElement, CurrentWrapper) => (
-      <CurrentWrapper routeId={id}>
+      <CurrentWrapper routeId={id} routeExports={routeExports}>
         {preElement}
       </CurrentWrapper>
     ), props.children);