+++ packages/runtime/src/routes.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { defer } from 'react-router-dom';
+import { defer, useRouteError } from 'react-router-dom';
 import type { RouteItem, RouteModules, RenderMode, DataLoaderConfig, RequestContext, ComponentModule } from './types.js';
 import RouteWrapper from './RouteWrapper.js';
 import { useAppContext } from './AppContext.js';
@@ -54,13 +54,13 @@ export async function loadRouteModules(routes: RouteModule[], originRouteModules
 export function WrapRouteComponent(options: {
   routeId: string;
   isLayout?: boolean;
-  RouteComponent: React.ComponentType;
+  routeExports: ComponentModule;
 }) {
-  const { routeId, isLayout, RouteComponent } = options;
+  const { routeId, isLayout, routeExports } = options;
   const { RouteWrappers } = useAppContext();
   return (
-    <RouteWrapper id={routeId} isLayout={isLayout} wrappers={RouteWrappers}>
-      <RouteComponent />
+    <RouteWrapper routeExports={routeExports} id={routeId} isLayout={isLayout} wrappers={RouteWrappers}>
+      <routeExports.default />
     </RouteWrapper>
   );
 }
@@ -80,6 +80,15 @@ export function RouteComponent({ id }: { id: string }) {
   return <Component />;
 }
 
+export function RouteErrorComponent() {
+  const error = useRouteError();
+  if (error) {
+    // Re-throws the error so it can be caught by App Error Boundary.
+    throw error;
+  }
+  return <></>;
+}
+
 /**
  * Create loader function for route module.
  */
@@ -90,7 +99,7 @@ interface LoaderData {
 
 export interface RouteLoaderOptions {
   routeId: string;
-  requestContext: RequestContext;
+  requestContext?: RequestContext;
   module: ComponentModule;
   renderMode: RenderMode;
 }