+++ packages/runtime/src/types.ts
@@ -1,5 +1,5 @@
 import type { IncomingMessage, ServerResponse } from 'http';
-import type { InitialEntry, AgnosticRouteObject, Router, Location } from '@remix-run/router';
+import type { InitialEntry, AgnosticRouteObject, Location } from '@remix-run/router';
 import type { ComponentType, PropsWithChildren } from 'react';
 import type { HydrationOptions, Root } from 'react-dom/client';
 import type { Params, RouteObject } from 'react-router-dom';
@@ -135,7 +135,8 @@ export type ComponentModule = {
 
 export type RouteItem = AgnosticRouteObject & {
   componentName: string;
-  exports: string[];
+  Component?: ComponentType<any>;
+  exports?: string[];
   layout?: boolean;
   children?: RouteItem[];
 };
@@ -221,9 +222,10 @@ export interface RuntimeModules {
 
 export interface AppRouterProps {
   routes?: RouteObject[];
-  router?: Router;
   routerContext?: any;
   location?: Location;
+  Component?: ComponentType<any>;
+  loaderData?: LoaderData;
 }
 
 export interface AppRouteProps {