+++ packages/plugin-rax-compat/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.0
+
+### Minor Changes
+
+- 2debc7f7: fix: appear event should not work on component
+- Updated dependencies [ee4141d5]
+  - rax-compat@0.2.0
+  - @ice/bundles@0.1.7
+
 ## 0.1.5
 
 ### Patch Changes